Refactor Services domain API specifications
