+++ package.json
@@ -51,6 +51,7 @@
     "generate:api:pagopaproxy": "rimraf generated/pagopa-proxy && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-pagopa-proxy/v1.6.0/api-spec/api-for-io.yaml --no-strict --out-dir generated/pagopa-proxy --request-types --response-decoders --client",
     "generate:api:lollipop": "rimraf generated/lollipop-api && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-functions-lollipop/724c3ba79d272d4e2555dbc2a8e1657bec9f84ce/openapi/internal.yaml --no-strict --out-dir generated/lollipop-api --request-types --response-decoders --client",
     "generate:api:services-app-backend": "rimraf generated/services-app-backend && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-services-cms/d2afbdf73e5b4ffd24aa46b025729ff2a9ed3cce/apps/app-backend/api/internal.yaml --no-strict --out-dir generated/services-app-backend --request-types --response-decoders --client",
+    "generate:proxy:services": "rimraf generated/services && gen-api-models --api-spec ./openapi/generated/api_services.yaml --no-strict --out-dir generated/services --request-types --response-decoders --client",
     "generate:proxy:eucovidcert-models": "rimraf generated/eucovidcert && gen-api-models --api-spec api_eucovidcert.yaml --out-dir generated/eucovidcert",
     "generate:proxy:third-party-service": "rimraf generated/third-party-service && gen-api-models --api-spec openapi/consumed/api-third-party.yaml --out-dir generated/third-party-service --request-types --response-decoders --client",
     "generate:proxy:remote-content-service": "rimraf generated/remote-content-service && gen-api-models --api-spec openapi/consumed/api_remote_content.yaml --out-dir generated/remote-content-service --request-types --response-decoders --client",
